refactor: GroupChat以外のクラスをModelsに移動
