+++ src/Command/Domain/Events/GroupChatCreated.php
@@ -2,10 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Akinoriakatsuka\CqrsEsExamplePhp\Command\Domain;
+namespace Akinoriakatsuka\CqrsEsExamplePhp\Command\Domain\Events;
 
 use J5ik2o\EventStoreAdapterPhp\Event;
 use DateTimeImmutable;
+
+use Akinoriakatsuka\CqrsEsExamplePhp\Command\Domain\Models\GroupChatId;
 
 interface GroupChatEvent extends Event {
     public function getAggregateId(): GroupChatId;
